backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
textstat==0.7.3
spacy==3.7.2
beautifulsoup4==4.12.2
Pillow==10.1.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
redis==5.0.1
reportlab==4.0.7
requests==2.31.0
cssselect==1.2.0
lxml==4.9.3
opencv-python==4.8.1.78
wcag-contrast-ratio==0.9
colour==0.1.5
tinycss2==1.2.1
pydantic-settings==2.1.0
nltk==3.8.1